Check parent paths for `package.json` file when checking if a module is ESM (#16)
This updates the `package.json` resolution algorithm to check for parent
paths of a given package specifier. For example, if we try to resolve
the closest `package.json` to `my-module/foo/bar/index.js`, previously
we would only check `my-module/foo/bar`. After this change, we also
check `my-module/foo`, in order from deep to shallow.