Add `@ts-bridge/resolver` package (#20)
This adds a new package, which is an implementation of Node.js' module
resolution algorithm. This can be used by the TS Bridge CLI to detect if
a module resolves, and whether it is CommonJS, ESM, or something else.

The aim of this package is to 100% match the implementation of Node.js.