Implement custom resolver in CLI (#24)
This is a significant refactor of the module resolution logic in the TS
Bridge CLI, in an attempt to get the logic closer to Node.js and other
projects. I've implemented `@ts-bridge/resolver` in the CLI and removed
all the existing custom resolver logic.

In addition, I rewrote a lot of tests to use the real file system rather
than a mocked file system. The mocked file system had several issues,
and it became unclear what was actually being tested. Now, most tests
use one of the fixtures in `test-utils/test/fixtures`, which are built
only once per `describe` block. Most tests are simply snapshot tests,
ensuring that the built files are as expected.

This replaces #23, as I've fixed the original bug in the new resolver
logic.