Inline shims instead of importing from `@ts-bridge/shims` (#35)
This removes the need for installing the `@ts-bridge/shims` package by
inlining shims instead, after @FrederikBolding and @rekmarks suggested
that this is the preferable solution.

Shims will be inserted in each file that needs them. For example, if a
file uses `__dirname`, which will be compiled to ESM, the `dirname`
helper function (and some dependency functions) are inserted into that
file. Note that shims are now opt-out (by specifying `--no-shims` or
`--shims false`) rather than opt-in (by installing the shims package
previously). I suppose this could be considered a breaking change.