Build project references in parallel (#56)
We can improve build performance by building project references in
parallel where possible. In this PR I've implemented an algorithm that
creates a worker thread for every project reference assuming all of its
dependencies are built, and the maximum number of workers isn't reached.

For example, given the following dependencies:

- Project A, depends on B, C.
- Project B, depends on C.
- Project C, no dependencies.
- Project D, no dependencies.

If the maximum number of workers (depends on CPU) allows for at least 2
workers, C and D will be built in parallel. Once C is built, B will be
built, and once that's done, A will be built. After some testing in some
`MetaMask/core` and `MetaMask/snaps`, this can reduce build times by 50%
(on my Macbook with M2 Max).

To reuse code where possible, I've moved a bunch of build code to a new
file `shared.ts` and the original `build.ts` now only contains code to
handle the build command.