Add end-to-end tests (#61)
This adds end-to-end tests for the build functionality. It covers Node
10 and Node 16 module resolution, as well as a project with references,
and a slightly more complex dependency graph than the current projects.
To accomplish this, I've added a simple command runner class which runs
the CLI in a separate process and listens to `stdout` and `stderr`.

The end-to-end tests run on the dist files, so building the project is
required for it to work. I've added a step in CI that waits for the
build step to complete, and uses the cached dist files from there.

The new project references fixture uses both Node 10 and Node 16 module
resolution, ensuring that they can be combined.

Closes #60.