Keep default imports when removing type imports (#78)
This fixes a bug where a default import would be removed if all named
imports are type imports. For example, in the following case:

```
import foo, { type Bar, type Baz } from 'module';
```

the entire import statement would be removed. After this change, the
default import is kept as-is and only the named imports are removed.

Closes #75.